added experimental interface to tags, problems, questions & solutions.
Added functions to access tags of sets defined in modules, problems of a gset identified by its tag, and solutions and challenges of those problems, identified by the gset tag.
These functions may largely come in handy for automated tests and debugging or getting to know the code-base, but probably not for actual challenge selection.
I may also need to re-organise the libraries, as the functions are not logically organized in their definition in their current form.