Library: rename hashmap hashtree, pass properties.
WARNING: this commit breaks lounge functionality due to changes in the hashmap
format. This is fixed in 756d9e6.

The change of name from crownset-hashmap to make-hashtree is due to
make-hashtree not necessarily having to be executed on crownsets, and due to a
clarification in terminology: a hashmap is a list containing hashtrees. A
hashtree is a pair with a hash as its car and '() or a list of further
hashtrees as its cdr.

* guilecraft/library-store.scm (crownset-hash-index): improve docstring.
  (hashtree?): new predicate.
  (crownset-hashmap): rename to 'make-hashtree', use 'rootset?' predicate,
  make car of hashtrees (prev. hashmaps) a pair of form '(hash . properties).
  (set-fullhash): use 'rootset?' predicate.
* guilecraft/module-server.scm (hashmap-provider): rename call to
  'make-hashtree' from 'crownset-hashmap'.
* tests/library-store.scm: new tests file.